LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_allow_top_navigation_by_user_activation_without_user_gesture-expected.txt


FAIL iframe_sandbox_allow_top_navigation_by_user_activation_without_user_gesture assert_equals: No beforeunload event should have been fired in the parent expected false but got true
